+++ NEWS.md
@@ -6,7 +6,7 @@
 * new `RawDatasetConnector` and `RelationalDatasetConnector` to execute get data from remote connections.
 * new wrapper functions to manipulate `DatasetConnector` and `Dataset` such as `get_dataset`, `load_dataset`, `as_relational`.
 
-* Argument `filter` in the `init` function was removed and the new argument `filter_states` was added. You can no longer modify the `app$datasets`, but must instead use this argument. 
+* You can no longer modify the `app$datasets`, but must instead use argument `filter` in the `init` function. 
 * New modules were created to create a module of nested teal modules, then another one that adds the right filter pane to each tab. The `teal::init` function stays unchanged.
 * The `teal::init` function now returns a `UI` function with an optional `id` argument. This allows to embed it into other applications. A split view of two teal applications side-by-side is one such example and shown in a vignette. `teal::init` was turned into a wrapper function around `module_teal_with_splash.R` and developers that want to embed teal as a Shiny module should directly work with these functions (`ui_teal_with_splash` and `srv_teal_with_splash`) instead of `teal::init`.
 * The `teal::init` function now has a title parameter to set the title of the browser window.